fix(Character): fixed windows compilation, needed explicit conversion

+++ src/game/objects/Character.cpp
@@ -73,12 +73,12 @@ void Character::setCollider() noexcept
     Vector3T<float> sca(this->getTransform().getScale());
 
     _collider.getBoundingBox().setBoundingBox(
-        {pos._x - sca._x * TILESIZE * 0.5,
+        {(float)(pos._x - sca._x * (float)TILESIZE * 0.5),
         pos._y,
-        pos._z - sca._z * TILESIZE * 0.5},
-        {pos._x + sca._x * TILESIZE * 0.5,
+        (float)(pos._z - sca._z * (float)TILESIZE * 0.5)},
+        {(float)(pos._x + sca._x * (float)TILESIZE * 0.5),
         pos._y,
-        pos._z + sca._z * TILESIZE * 0.5});
+        (float)(pos._z + sca._z * (float)TILESIZE * 0.5)});
 }
 
 void Character::setState(const int &state) noexcept
